+++ scripts/JSRootCore.js
@@ -89,7 +89,7 @@
    }
 } (function(JSROOT) {
 
-   JSROOT.version = "dev 29/05/2017";
+   JSROOT.version = "dev 12/06/2017";
 
    JSROOT.source_dir = "";
    JSROOT.source_min = false;